fix: `interface_statement` have `decorator_list`
Indeed, the web documentation makes it appear at first glance that interface does not have decorators,
but looking at the implementation, it does.
It is also shown in the example code.

<https://github.com/microsoft/typespec/blob/a939c140980805dc66a9b6fef70be5647b1c9680/packages/samples/specs/versioning/main.tsp#L51>

The PR to fix the standard and documentation has not yet been merged, but has already been submitted.

[docs: interface statement can have decorators by ncaq · Pull Request #5202 · microsoft/typespec](https://github.com/microsoft/typespec/pull/5202)

I found it because of a problem that the parser here could not highlight the sample code correctly.

I'll submit a PR here first as well.