build: add `package-lock.json`
`package-lock.json` is a file created by npm.
The npm can skip some steps and install dependencies faster because it knows exactly which versions to use.
It locks the versions of dependencies, reducing the risk of accidentally using a version with known security issues.

Since we have tree-sitter and Makefile,
sometimes I want to blow through the cache and proceed to `npm install`,
but then it's hard if `npm install` is slow.
I would therefore like to add this.